Add new users to the db

+++ app/src/main/java/com/example/sdgbachelorproject/model/data/HeatingResultItem.kt
@@ -2,7 +2,7 @@ package com.example.sdgbachelorproject.model.data
 
 data class HeatingResultItem(
     val buildingType: Int,
-    val heatingConsumption: Int,
+    val heatingConsumption: Float,
     val id: Int,
     val numberOfRooms: Int,
     val temperatureInYourHoushold: Int,
